Update to BeanShell 2.1.0
This version of bsh is not available on Maven Central, unfortunately.
I guess the BeanShell maintainers are no longer publishing artifacts
there for some reason. Fortunately, the invesdwin organization has
published a bug-fixed version of BeanShell 2.1.0 to their repository,
which maven.scijava.org now proxies, so that we can use it! \^_^/

See also:
https://github.com/scijava/scripting-beanshell/issues/2#issuecomment-1013933143