Do not claim ownership of .java source files
BeanShell is a very close approximation to Java syntax. But it does not
handle everything you can write into a .java file. And it certainly does
not actually try to compile and run it with javac. It is wrong for
BeanShell to snarf up these files, and in particular it clashes with the
SciJava scripting-java component, which *does* actually use javac.